whisp/requirements.txt
python-dotenv
yt-dlp
pydub
pyannote-audio
openai-whisper
